v2 synced

+++ epan/dwarf.c
@@ -39,7 +39,7 @@ dissect_uleb128(tvbuff_t *tvb, gint offset, guint64 *value)
         byte = tvb_get_guint8(tvb, offset);
         offset += 1;
 
-        *value |= (byte & 0x7F) << shift;
+        *value |= ((guint64)(byte & 0x7F) << shift);
         shift += 7;
     } while (byte & 0x80);
 
@@ -59,12 +59,12 @@ dissect_leb128(tvbuff_t *tvb, gint offset, gint64 *value)
         byte = tvb_get_guint8(tvb, offset);
         offset += 1;
 
-        *value |= (byte & 0x7F) << shift;
+        *value |= ((guint64)(byte & 0x7F) << shift);
         shift += 7;
     } while (byte & 0x80);
 
     if (shift < 64 && byte & 0x40)
-        *value |= - (1 << shift);
+        *value |= - ((gint64)1 << shift);
 
     return offset - start_offset;
 }
